Added rudimentary automatic guithub build publishing script.
Also added github uploader script (http://blog.caurea.org/2011/01/10/github-upload-script.html)

The mkdist-osx.sh script will build Tundra for OSX, Win32 (via mingw cross compilation)
and the documentation PDF, package and rename everything with a moniker (e.g. a version tag)
specified on the command line and upload the three result files to github automatically.

+++ support/mkdist-osx.sh
@@ -1,26 +1,57 @@
 #! /bin/sh
 
+MONIKER=$1
+
+if [ x$MONIKER = x ]; then
+	echo "error: specify a moniker"
+	exit 1
+fi
+
+test -d dists || mkdir dists
+DISTOSX=tundra-$MONIKER-osxintel64
+DISTWIN=tundra-$MONIKER-win32
+DISTMAN=tundra-$MONIKER-manual
+
 test -f tundra.lua || exit 1
 
 find examples -name tundra-output -exec rm -rf {} \;
-find examples -name .tundra-\* -exec rm -f {} \;
+find . -name .tundra-\* -exec rm -f {} \;
 
-rm -rf build dist
+rm -rf build $DISTOSX $DISTWIN tundra-output
 
 mkdir build
 cd build
-cmake ..
-make
+cmake .. || exit 1
+make || exit 1
 cd ..
 
-TUNDRA_HOME=$PWD build/tundra standalone release macosx-clang
+TUNDRA_HOME=$PWD build/tundra standalone release macosx-clang || exit 1
+TUNDRA_HOME=$PWD build/tundra standalone release macosx-crosswin32 || exit 1
 
-mkdir dist
-mkdir dist/doc
-cp -r README.md COPYING examples dist
-cp doc/manual.asciidoc dist/doc
-cp tundra-output/macosx-clang-release-standalone/tundra dist
-git log -1 >> dist/SNAPSHOT_REVISION
-find dist -name \*.swp -exec rm {} \;
-find dist -name .DS_Store -exec rm {} \;
+mkdir $DISTOSX
+mkdir $DISTOSX/doc
+cp -r README.md COPYING examples $DISTOSX
+cp doc/manual.asciidoc $DISTOSX/doc
+git log -1 >> $DISTOSX/SNAPSHOT_REVISION
+find $DISTOSX -name \*.swp -exec rm {} \;
+find $DISTOSX -name .DS_Store -exec rm {} \;
 
+cp -r $DISTOSX $DISTWIN
+
+cp tundra-output/macosx-clang-release-standalone/tundra $DISTOSX || exit 1
+cp tundra-output/macosx-crosswin32-release-standalone/tundra.exe $DISTWIN || exit 1
+
+tar cjvf dists/$DISTOSX.tar.bz2 $DISTOSX || exit 1
+zip -r dists/$DISTWIN.zip $DISTWIN || exit 1
+
+rm -rf $DISTOSX $DISTWIN
+
+make -C doc
+cp doc/manual.pdf dists/$DISTMAN.pdf
+
+support/github-upload.rb dists/$DISTOSX.tar.bz2 || exit 1
+support/github-upload.rb dists/$DISTWIN.zip || exit 1
+support/github-upload.rb dists/$DISTMAN.pdf || exit 1
+
+echo
+echo "All done, tundra $MONIKER uploaded successfully"
